+++ src/com/pacman/renderer/DefaultRenderer.java
@@ -1,6 +1,7 @@
 package com.pacman.renderer;
 
-import org.lwjgl.util.Point;
+import java.awt.Point;
+
 import org.newdawn.slick.Graphics;
 
 public class DefaultRenderer implements Renderer {
@@ -8,10 +9,10 @@ public class DefaultRenderer implements Renderer {
 	@Override
 	public void render(Renderable renderable, Graphics g) {
 		Point position = renderable.getPosition();
-		draw(renderable, g, position.getX(), position.getY());
+		draw(renderable, g, (float) position.getX(), (float) position.getY());
 	}
 
-	private void draw(Renderable renderable, Graphics g, Integer x, Integer y) {
+	private void draw(Renderable renderable, Graphics g, Float x, Float y) {
 		g.pushTransform();
 		g.translate(x, y);
 		renderable.draw(g);
